path fix after moving files
