+++ tooling/cmd/sysmon-modular/command_diff.go
@@ -4,8 +4,8 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/olafhartong/sysmon-modular/internal/semantic"
-	"github.com/olafhartong/sysmon-modular/internal/sysmonxml"
+	"github.com/olafhartong/sysmon-modular/tooling/internal/semantic"
+	"github.com/olafhartong/sysmon-modular/tooling/internal/sysmonxml"
 )
 
 func runDiff(args []string) error {
