+++ tooling/go.mod
@@ -1,3 +1,3 @@
-module github.com/olafhartong/sysmon-modular
+module github.com/olafhartong/sysmon-modular/tooling
 
 go 1.22
