+++ tooling/internal/validate/validate_test.go
@@ -4,7 +4,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/olafhartong/sysmon-modular/internal/sysmonxml"
+	"github.com/olafhartong/sysmon-modular/tooling/internal/sysmonxml"
 )
 
 func TestSchemaFlagsUnknownEventAndInvalidOnmatch(t *testing.T) {
